Combine initialize, autonomous, and opcontrol into main
